Move client into tools
