+++ bench/exo_eval.py
@@ -42,9 +42,8 @@ from pathlib import Path
 from typing import Any
 
 import httpx
-from exo.client import (
-    ExoClient,
-    ExoHttpError,
+from exo_tools.client import ExoClient, ExoHttpError
+from exo_tools.harness import (
     add_common_instance_args,
     capture_cluster_snapshot,
     find_existing_instance,
